WIP - fix mdfc
This aims to fix mdfcs and transform cards not having the correct hash. It looks up the card in a dict and it then gets the cockatrice name for it. This means that players can have the full name of a mdfc or transform card and cockatrice will still have the correct hash as this program shows.